+++ NEWS.md
@@ -1,5 +1,6 @@
 # fluxible 0.1.2
 
+* Added option in `flux_gep` to keep all columns without naming them.
 * Order of columns in the output of flux_match is more logical (f_start before f_end...).
 
 # fluxible 0.1.1